renamed data-service.ts to data-api.ts for consistency
