+++ src/lib/core/__test__/hooks/useSession.test.tsx
@@ -3,7 +3,7 @@ import { act, renderHook } from '@testing-library/react-hooks';
 import { useSession, Status } from '../../hooks/session';
 import { Session, NewSession } from '../../types/session';
 import { SessionClient } from '../../api/session-api';
-import { DataClient } from '../../api/data-service';
+import { DataClient } from '../../api/data-api';
 import {
   StudyMetadata,
   StudyRecord,
