+++ src/lib/core/components/EDASessionListContainer.tsx
@@ -3,7 +3,7 @@ import { useWdkStudyRecord, useStudyMetadata } from '../hooks/study';
 import { LoadError } from '@veupathdb/wdk-client/lib/Components';
 import { SessionClient } from '../api/session-api';
 import { SubsettingClient } from '../api/subsetting-api';
-import { DataClient } from '../api/data-service';
+import { DataClient } from '../api/data-api';
 import { WorkspaceContext } from '../context/WorkspaceContext';
 
 interface Props {
