+++ src/lib/core/context/WorkspaceContext.ts
@@ -1,6 +1,6 @@
 import { createContext } from 'react';
 import { SubsettingClient } from '../api/subsetting-api';
-import { DataClient } from '../api/data-service';
+import { DataClient } from '../api/data-api';
 import { SessionClient } from '../api/session-api';
 import { StudyMetadata, StudyRecord, StudyRecordClass } from '../types/study';
