+++ src/lib/core/index.ts
@@ -4,7 +4,7 @@ export * from './types/filter';
 export * from './hooks/session';
 export * from './hooks/workspace';
 export * from './api/subsetting-api';
-export * from './api/data-service';
+export * from './api/data-api';
 export * from './api/session-api';
 export * from './context/WorkspaceContext';
 export { EDAWorkspaceContainer } from './components/EDAWorkspaceContainer';
